pyproject.toml
[tool.poetry]
name = "elasticsearch-haystack"
version = "0.1.0"
description = ""
authors = ["Robin Winslow Morris <robin@robinwinslow.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.3"
elasticsearch = "^8.9.0"
haystack-ai = "^0.9.0"
tox = "^4.11.0"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
